+++ tapo/examples/tapo_h100.rs
@@ -2,7 +2,7 @@
 use std::env;
 
 use log::{info, LevelFilter};
-use tapo::{responses::ChildDeviceResult, ApiClient};
+use tapo::{responses::ChildDeviceHubResult, ApiClient};
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -31,7 +31,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     for child in child_device_list {
         match child {
-            ChildDeviceResult::KE100(device) => {
+            ChildDeviceHubResult::KE100(device) => {
                 info!(
                     "Found KE100 child device with nickname: {}, id: {}, current temperature: {} {:?} and target temperature: {} {:?}.",
                     device.nickname,
@@ -42,7 +42,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                     device.temperature_unit,
                 );
             }
-            ChildDeviceResult::S200B(device) => {
+            ChildDeviceHubResult::S200B(device) => {
                 let s200b = hub.s200b(&device.device_id);
                 let trigger_logs = s200b.get_trigger_logs(5, 0).await?;
 
@@ -51,7 +51,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                     device.nickname, device.device_id, trigger_logs
                 );
             }
-            ChildDeviceResult::T100(device) => {
+            ChildDeviceHubResult::T100(device) => {
                 let t100 = hub.t100(&device.device_id);
                 let trigger_logs = t100.get_trigger_logs(5, 0).await?;
 
@@ -60,7 +60,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                     device.nickname, device.device_id, device.detected, trigger_logs
                 );
             }
-            ChildDeviceResult::T110(device) => {
+            ChildDeviceHubResult::T110(device) => {
                 let t110 = hub.t110(&device.device_id);
                 let trigger_logs = t110.get_trigger_logs(5, 0).await?;
 
@@ -69,7 +69,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                     device.nickname, device.device_id, device.open, trigger_logs
                 );
             }
-            ChildDeviceResult::T300(device) => {
+            ChildDeviceHubResult::T300(device) => {
                 let t300 = hub.t300(&device.device_id);
                 let trigger_logs = t300.get_trigger_logs(5, 0).await?;
 
@@ -82,7 +82,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                     trigger_logs
                 );
             }
-            ChildDeviceResult::T310(device) | ChildDeviceResult::T315(device) => {
+            ChildDeviceHubResult::T310(device) | ChildDeviceHubResult::T315(device) => {
                 let t31x = hub.t315(&device.device_id);
                 let temperature_humidity_records = t31x.get_temperature_humidity_records().await?;