+++ tapo/src/api.rs
@@ -8,6 +8,7 @@ mod hub_handler;
 mod light_handler;
 mod plug_energy_monitoring_handler;
 mod plug_handler;
+mod power_strip_handler;
 mod protocol;
 
 pub use api_client::*;
@@ -20,3 +21,4 @@ pub use hub_handler::*;
 pub use light_handler::*;
 pub use plug_energy_monitoring_handler::*;
 pub use plug_handler::*;
+pub use power_strip_handler::*;